+++ bench/utils.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace bench
+{
+std::string expensive(std::size_t);
+}