+++ include/threadpool.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <iostream>
+
 #include <algorithm>
 #include <condition_variable>
 #include <functional>
@@ -17,14 +19,18 @@
   do                                                                           \
   {                                                                            \
     if (pool->hooks)                                                           \
+    {                                                                          \
       pool->hooks->HOOK();                                                     \
+    }                                                                          \
   } while (0)
 
 #define CALL_HOOK_POOL(HOOK)                                                   \
   do                                                                           \
   {                                                                            \
     if (hooks)                                                                 \
+    {                                                                          \
       hooks->HOOK();                                                           \
+    }                                                                          \
   } while (0)
 
 namespace ThreadPool
@@ -43,20 +49,23 @@ public:
    *
    *  The number of workers will be deduced from hardware.
    */
-  ThreadPool();
+  explicit ThreadPool();
 
   /*! \brief Constructs a ThreadPool.
    *  \param pool_size Number of threads to start.
    */
-  ThreadPool(std::size_t pool_size);
+  explicit ThreadPool(std::size_t pool_size);
+
+  /*! \brief Constructs a ThreadPool.
+   *  \param hooks Hooks to register in the pool.
+   */
+  explicit ThreadPool(std::shared_ptr<Hooks> hooks);
 
   /*! \brief Constructs a ThreadPool.
    *  \param pool_size Number of threads to start.
-   *  \param max_pool_size Maximum number of threads allowed, this will be used
-   *  by the pool to extend the number of threads temporarily when all threads
-   *  are used.
+   *  \param hooks Hooks to register in the pool.
    */
-  ThreadPool(std::size_t pool_size, std::size_t max_pool_size);
+  ThreadPool(std::size_t pool_size, std::shared_ptr<Hooks> hooks);
 
   /*! \brief Stops the pool and clean all workers.
    */
@@ -74,9 +83,9 @@ public:
   auto run(Function&& f, Args&&... args)
     -> std::future<typename std::result_of<Function(Args...)>::type>;
 
-  /*! \brief Stop the SingleQueue.
+  /*! \brief Stop the ThreadPool.
    *
-   * A stopped SingleQueue will discard any task dispatched to it. All workers
+   * A stopped ThreadPool will discard any task dispatched to it. All workers
    * will discard new tasks, but the threads will not exit.
    */
   void stop();
@@ -100,9 +109,9 @@ public:
   void register_hooks(std::shared_ptr<Hooks> hooks);
 
   /*! \brief Check the state of the threadpool
-   *  \returns True if the bool is stopped, false otherwise.
+   *  \returns True if the ThreadPool is stopped, false otherwise.
    */
-  bool is_stopped() const;
+  bool is_stopped() const noexcept;
 
   /*! \brief Check on the number of threads not currently working.
    *  \returns The number of threads currently waiting for a task.
@@ -110,7 +119,7 @@ public:
    * The number might be imprecise, as between the time the value is read and
    * returned, a thread might become unavailable.
    */
-  std::size_t threads_available() const;
+  std::size_t threads_available() const noexcept;
 
   /*! \brief Check on the number of threads currently working.
    *  \returns The number of threads currently working.
@@ -118,23 +127,34 @@ public:
    * The number might be imprecise, as between the time the value is read and
    * returned, a thread might finish a task and become available.
    */
-  std::size_t threads_working() const;
+  std::size_t threads_working() const noexcept;
 
 private:
+  using task_type = std::packaged_task<void()>;
+
   /*! \brief Starts the pool when the pool is constructed.
    *
    *  It will starts _pool_size threads.
    */
   void start_pool();
 
-  /*! \brief Joins all threads in the pool.
+  /*! \brief Clean the pool and join threads of dead workers.
    *
-   *  Should only be called from destructor.
+   *  This method may be called at any time by any thread putting a job in the
+   *  queue. This function acquires a lock on the pool vector.
    */
   void clean();
 
+  /*! \brief Joins all threads in the pool.
+   *
+   *  Should only be called from destructor. This method will stop all the
+   *  worker and join the corresponding thread.
+   */
+  void terminate();
+
   /*! \brief Find the worker for which to dispatch the tasks
-   *  \return The index in the worker array to which a task should be dispatch.
+   *  \return The index in the worker array to which a task should be
+   * dispatch.
    */
   std::size_t get_dispatch_worker();
 
@@ -156,24 +176,37 @@ private:
     /*! \brief Construct a worker.
      *  \param pool The ThreadPool the worker works for.
      */
-    Worker(ThreadPool* pool);
+    explicit Worker(ThreadPool* pool);
 
     /*! \brief Poll task from the queue.
      *  \param nb_task Number of tasks to run and then exit. If 0 then run until
      *  the ThreadPool stops.
      */
-    void operator()(std::size_t nb_task);
+    void operator()();
+
+    /*! \brief Stop the Worker
+     */
+    void stop();
+
+    /*! \brief Start the Worker
+     */
+    void start();
+
+    /*! \brief Check the state of the worker
+     *  \returns True if the worker is stopped, false otherwise.
+     */
+    bool is_stopped() const noexcept;
 
     /*! \brief The task queue.
      *
      *  Access to this task queue should **always** be done while locking using
      *  tasks_lock mutex.
      */
-    std::queue<std::packaged_task<void()>> tasks;
+    std::queue<task_type> tasks;
 
-    /*! \brief Mutex regulating acces to _tasks.
+    /*! \brief Mutex regulating acces to tasks.
      */
-    std::mutex tasks_lock;
+    mutable std::mutex tasks_lock;
 
     /*! \brief Condition variable used to wake the worker for when a task is
      *  available.
@@ -183,25 +216,24 @@ private:
     std::atomic<std::size_t> queue_size;
 
   private:
+    task_type extract_task(std::queue<task_type>& task_queue);
+    std::pair<bool, task_type> work_steal();
+    std::pair<bool, task_type> find_task();
+    void wait_for_start();
+
     /*! \brief Captured ThreadPool that the worker works for.
      */
     ThreadPool* pool;
-  };
 
-  /*! \brief Start a worker for a nb_task.
-   *  \param nb_task Number of tasks to execute.
-   *
-   *  If nb_task is 0 (default arg) the worker will remain alive until
-   *  threadpool is stopped.
-   */
-  void add_worker(std::size_t nb_task = 0);
+    std::atomic<bool> stopped;
+    std::atomic<bool> started;
+  };
 
   /*! \brief Check if the pool can spawn more workers, and spawn one for a
    *  single task
    *
    *  It will check the current number of spawned threads and if it can spawn
-   *  or not a new thread. If a thread can be spawned, one is created for a
-   *  single task.
+   *  or not a new thread. If a thread can be spawned, one is created.
    */
   void check_spawn_single_worker();
 
@@ -209,11 +241,11 @@ private:
    *
    *  Emplacing in this vector construct and launch a thread.
    */
-  std::vector<std::thread> pool;
+  std::vector<std::pair<std::thread, std::unique_ptr<Worker>>> pool;
 
-  /*! \brief Vector of the workers in the pool.
+  /*! \brief Mutex regulating acces to the pool
    */
-  std::vector<std::unique_ptr<Worker>> workers;
+  mutable std::mutex pool_lock;
 
   /*! \brief Number of waiting threads in the pool.
    */
@@ -223,16 +255,6 @@ private:
    */
   std::atomic<std::size_t> working_threads;
 
-  /*! \brief Size of the pool.
-   */
-  const std::size_t pool_size;
-
-  /*! \brief Max possible size of the pool.
-   *
-   *  This parameter is used to add additional threads if
-   */
-  const std::size_t max_pool_size;
-
   /*! \brief Boolean representing if the pool is stopped.
    *
    */
@@ -241,29 +263,36 @@ private:
   /*! \brief Struct containing all hooks the threadpool will call.
    */
   std::shared_ptr<Hooks> hooks;
+
+  /*! \brief Size of the pool.
+   */
+  const std::size_t pool_size;
 };
 
 // ThreadPool implementation
-// public:
 
 inline ThreadPool::ThreadPool()
-  : ThreadPool(std::thread::hardware_concurrency(),
-               std::thread::hardware_concurrency())
+  : ThreadPool(std::thread::hardware_concurrency(), nullptr)
 {
 }
 
 inline ThreadPool::ThreadPool(std::size_t pool_size)
-  : ThreadPool(pool_size, pool_size)
+  : ThreadPool(pool_size, nullptr)
 {
 }
 
-inline ThreadPool::ThreadPool(std::size_t pool_size, std::size_t max_pool_size)
+inline ThreadPool::ThreadPool(std::shared_ptr<Hooks> hooks)
+  : ThreadPool(std::thread::hardware_concurrency(), hooks)
+{
+}
+
+inline ThreadPool::ThreadPool(std::size_t pool_size,
+                              std::shared_ptr<Hooks> hooks)
   : waiting_threads(0)
   , working_threads(0)
-  , pool_size(pool_size)
-  , max_pool_size(max_pool_size)
   , stopped(false)
-  , hooks(nullptr)
+  , hooks(hooks)
+  , pool_size(pool_size)
 {
   start_pool();
 }
@@ -271,7 +300,7 @@ inline ThreadPool::ThreadPool(std::size_t pool_size, std::size_t max_pool_size)
 inline ThreadPool::~ThreadPool()
 {
   stop();
-  clean();
+  terminate();
 }
 
 template <typename Function, typename... Args>
@@ -282,174 +311,191 @@ auto ThreadPool::run(Function&& f, Args&&... args)
 
   // Create a packaged task from the callable object to fetch its result
   // with get_future()
-  auto inner_task = std::packaged_task<task_return_type()>(
+  auto task = std::packaged_task<task_return_type()>(
     std::bind(std::forward<Function&&>(f), std::forward<Args&&...>(args)...));
+  auto result = task.get_future();
 
   std::size_t idx = this->get_dispatch_worker();
 
-  // TODO: If the pool can and should spawn more workers, spawn one for a
-  // single task. The problem here is to handle synchronisation of modifying
-  // the workers vector (when the threadpool is used by multiple thread. If
-  // only one thread handles it, there is no problem.
-  //
-  // check_spawn_single_worker();
-
-  auto result = inner_task.get_future();
   if (stopped)
   {
     return result;
   }
-  dispatch_work(idx, std::move(inner_task));
+
+  dispatch_work(idx, std::move(task));
   return result;
 }
 
-inline void ThreadPool::stop()
+inline std::size_t ThreadPool::get_dispatch_worker()
 {
-  stopped = true;
-  for (auto& w : workers)
-  {
-    // This is needed in order for the worker to rerun the lambda in the
-    // condition variable to check the condition (stopped) and exit the wait.
-    // If this is not here, sometime a deadlock can occur as the thread will not
-    // check the stopped boolean and remain waiting, while the main thread try
-    // to join it and stops.
-    {
-      std::unique_lock<decltype(Worker::tasks_lock)> lock(w->tasks_lock);
-    }
-    w->cv_variable.notify_all();
-  }
-}
-
-// ThreadPool implementation
-// private:
-
-inline void ThreadPool::start_pool()
-{
-  for (std::size_t i = 0; i < pool_size; i++)
-  {
-    add_worker();
-  }
-}
-
-inline void ThreadPool::clean()
-{
-  for (auto& t : pool)
-  {
-    if (t.joinable())
-    {
-      CALL_HOOK_POOL(on_worker_die);
-      t.join();
-    }
-  }
-}
-
-inline void ThreadPool::add_worker(std::size_t nb_task)
-{
-  auto w = std::unique_ptr<Worker>(new Worker(this));
-  pool.emplace_back(std::ref(*w), nb_task);
-  workers.push_back(std::move(w));
-}
-
-inline void ThreadPool::check_spawn_single_worker()
-{
-  // Check if we are allowed to spawn a worker
-  if (max_pool_size > pool_size)
-  {
-    // Check if we have space to spawn a worker, and if it is valuable.
-    if (working_threads.load() + waiting_threads.load() < max_pool_size)
-    {
-      CALL_HOOK_POOL(on_worker_add);
-      add_worker(1);
-    }
-  }
+  // For now we dispatch at random. If the random index is on a worker that is
+  // stopped but now yet cleaned, return another idx;
+  static std::random_device seeder;
+  static std::mt19937 engine(seeder());
+  std::uniform_int_distribution<int> dist(0, pool.size() - 1);
+  return dist(engine);
 }
 
 template <typename TaskType>
 inline void ThreadPool::dispatch_work(const std::size_t idx, TaskType task)
 {
-  workers[idx]->queue_size++;
+  auto& worker = pool[idx];
   {
-    std::lock_guard<decltype(Worker::tasks_lock)> lk(workers[idx]->tasks_lock);
-    workers[idx]->tasks.emplace(std::move(task));
+    std::lock_guard<decltype(Worker::tasks_lock)> lk(worker.second->tasks_lock);
+    worker.second->tasks.emplace(std::move(task));
   }
-  workers[idx]->cv_variable.notify_one();
+  worker.second->queue_size--;
+  worker.second->cv_variable.notify_one();
 }
-
-inline std::size_t ThreadPool::get_dispatch_worker()
-{
-  // Random implementation
-  static std::random_device seeder;
-  static std::mt19937 engine(seeder());
-  static std::uniform_int_distribution<int> dist(0, workers.size() - 1);
-  return dist(engine);
-  /*std::size_t idx = 0;
-  std::size_t it = 0;
-  std::size_t min = 0;
-  std::for_each(std::begin(workers), std::end(workers),
-                [&it, &idx, &min](const std::unique_ptr<Worker>& w) {
-                  if (min < w->queue_size)
-                  {
-                    idx = it;
-                    min = w->queue_size;
-                  }
-                  it++;
-                });
-  return idx;*/
-}
-
-// Worker implementation
-// public:
 
 inline ThreadPool::Worker::Worker(ThreadPool* pool)
   : pool(pool)
+  , stopped(false)
+  , started(false)
 {
 }
 
-inline void ThreadPool::Worker::operator()(std::size_t nb_task)
+inline void ThreadPool::Worker::operator()()
 {
-  for (std::size_t i = 0; i != nb_task || nb_task == 0; i++)
+  wait_for_start();
+  for (;;)
   {
     // Thread is waiting
     pool->waiting_threads += 1;
 
-    std::unique_lock<decltype(tasks_lock)> lock(tasks_lock);
-    cv_variable.wait(lock, [&] { return pool->stopped || !tasks.empty(); });
-
-    // Pool is stopped, discard task and exit
-    if (pool->stopped)
-      return;
+    auto task = find_task();
+    if (!task.first)
+    {
+      break;
+    }
 
     CALL_HOOK_WORKER(pre_task_hook);
 
     pool->waiting_threads -= 1;
     pool->working_threads += 1;
 
-    // Fetch task
-    std::packaged_task<void()> task = std::move(tasks.front());
-    tasks.pop();
-    queue_size--;
+    task.second();
 
-    // Release lock and exec task
-    lock.unlock();
-    task();
-
-    // Task done
     CALL_HOOK_WORKER(post_task_hook);
     pool->working_threads -= 1;
   }
 }
 
-inline bool ThreadPool::is_stopped() const
+inline void ThreadPool::Worker::wait_for_start()
+{
+  std::unique_lock<decltype(tasks_lock)> lock(tasks_lock);
+  cv_variable.wait(lock, [&] { return this->started == true; });
+}
+
+inline std::pair<bool, ThreadPool::task_type> ThreadPool::Worker::find_task()
+{
+  for (;;)
+  {
+    /*if (queue_size == 0)
+    {
+      // Try work stealing
+      auto res = std::move(work_steal());
+      if (res.first)
+      {
+        return res;
+      }
+    }*/
+
+    // Wait for a task in our queue
+    std::unique_lock<decltype(tasks_lock)> lock(tasks_lock);
+    cv_variable.wait(
+      lock, [&] { return pool->stopped || this->stopped || !tasks.empty(); });
+
+    // Pool is stopped
+    if (pool->stopped || this->stopped)
+    {
+      return std::pair<bool, task_type>(false, task_type{});
+    }
+
+    return std::pair<bool, task_type>(true, extract_task(tasks));
+  }
+}
+
+inline std::pair<bool, ThreadPool::task_type> ThreadPool::Worker::work_steal()
+{
+  /*std::lock_guard<decltype(pool_lock)> lk(pool->pool_lock);
+  for (auto& w : pool->pool)
+  {
+    if (pool->stopped || this->stopped)
+    {
+      return std::pair<bool, task_type>(false, task_type{});
+    }
+
+    if (w.second->queue_size == 0)
+    {
+      continue;
+    }
+
+    // FIXME: add try lock
+    std::lock_guard<decltype(Worker::tasks_lock)> lk(w.second->tasks_lock);
+
+    return std::pair<bool, task_type>(true, extract_task(tasks));
+  }
+  return std::pair<bool, task_type>(false, task_type{});*/
+}
+
+inline ThreadPool::task_type
+ThreadPool::Worker::extract_task(std::queue<ThreadPool::task_type>& task_queue)
+{
+  task_type task = std::move(task_queue.front());
+  task_queue.pop();
+  queue_size--;
+  return task;
+}
+
+inline void ThreadPool::stop()
+{
+  stopped = true;
+  std::lock_guard<decltype(pool_lock)> pl(pool_lock);
+  for (const auto& w : pool)
+  {
+    w.second->stop();
+  }
+}
+
+inline void ThreadPool::terminate()
+{
+  std::lock_guard<decltype(pool_lock)> pl(pool_lock);
+  // Join everything
+  for (auto& w : pool)
+  {
+    w.first.join();
+    CALL_HOOK_POOL(on_worker_die);
+  }
+}
+
+inline void ThreadPool::start_pool()
+{
+  std::lock_guard<decltype(pool_lock)> pl(pool_lock);
+  for (std::size_t i = 0; i < pool_size; i++)
+  {
+    auto w = std::unique_ptr<Worker>(new Worker(this));
+    pool.emplace_back(std::ref(*w), std::move(w));
+    CALL_HOOK_POOL(on_worker_add);
+  }
+
+  for (auto& w : pool)
+  {
+    w.second->start();
+  }
+}
+
+inline bool ThreadPool::is_stopped() const noexcept
 {
   return stopped;
 }
 
-inline std::size_t ThreadPool::threads_available() const
+inline std::size_t ThreadPool::threads_available() const noexcept
 {
   return waiting_threads.load();
 }
 
-inline std::size_t ThreadPool::threads_working() const
+inline std::size_t ThreadPool::threads_working() const noexcept
 {
   return working_threads.load();
 }
@@ -458,4 +504,28 @@ inline void ThreadPool::register_hooks(std::shared_ptr<Hooks> hooks)
 {
   this->hooks = hooks;
 }
+
+inline void ThreadPool::Worker::stop()
+{
+  {
+    std::lock_guard<decltype(Worker::tasks_lock)> lock(tasks_lock);
+    stopped = true;
+  }
+  cv_variable.notify_one();
+}
+
+inline void ThreadPool::Worker::start()
+{
+  {
+    std::lock_guard<decltype(Worker::tasks_lock)> lock(tasks_lock);
+    started = true;
+  }
+  cv_variable.notify_one();
+}
+
+inline bool ThreadPool::Worker::is_stopped() const noexcept
+{
+  return stopped;
+}
+
 } // namespace ThreadPool