+++ tests/hooks_test.hpp
@@ -1,13 +1,11 @@
 #pragma once
 
-#include "threadpool.hpp"
-
-#include <chrono>
 #include <future>
 #include <gtest/gtest.h>
 #include <memory>
 #include <thread>
-#include <utility>
+
+#include "threadpool.hpp"
 
 struct DummyHooks : public ThreadPool::Hooks
 {