+++ tests/threadpool_test.hpp
@@ -1,12 +1,11 @@
 #pragma once
 
-#include "threadpool.hpp"
-#include <chrono>
 #include <future>
 #include <gtest/gtest.h>
 #include <memory>
 #include <thread>
-#include <utility>
+
+#include "threadpool.hpp"
 
 class ThreadPoolTest : public ::testing::Test
 {